creado el controlador y la ruta, ademas su prueba, miembro 1 del equipo

+++ database/migrations/2024_07_04_023858_create_unidads_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('unidads', function (Blueprint $table) {
             $table->id('idUnidad');
             $table->string('nombre');
-            $table->integer('codigo_presupuesto');
+            $table->unsignedBigInteger('codigo_presupuesto');
             $table->foreign('codigo_presupuesto')->references('codigoPresupuesto')->on('presupuestos')->onDelete('cascade');
             $table->timestamps();
         });
